[us-14844]  remove store management page on storefront
