--- src/@next/components/molecules/StoreTagTiles/index.ts
@@ -1,1 +1,0 @@
-export * from "./StoreTabTiles";
