--- src/@next/pages/StorePage/index.ts
@@ -1,1 +1,0 @@
-export * from "./StorePage";
