--- src/@next/pages/StorePage/styles.ts
@@ -1,6 +1,0 @@
-import { styled } from "@styles";
-
-export const Wrapper = styled.div`
-  display: flex;
-  flex-direction: column;
-`;
