+++ src/@next/pages/index.ts
@@ -4,4 +4,3 @@ export * from "./PasswordReset";
 export * from "./CartPage";
 export * from "./CheckoutPage";
 export * from "./ThankYouPage";
-export * from "./StorePage";
