--- src/pages/store/[id].ts
@@ -1,4 +1,0 @@
-import { StorePage, StorePageProps } from "@temp/views/StorePage";
-
-export default StorePage;
-StorePage.getInitialProps = async ({ query }) => ({ query } as StorePageProps);
