--- src/pages/store/index.ts
@@ -1,3 +1,0 @@
-import { StorePage } from "@pages/StorePage";
-
-export default StorePage;
