+++ src/paths.ts
@@ -36,7 +36,6 @@ export const paths = {
   // FIXME: User order should be accessible via order id
   accountOrderDetail: `${accountBaseUrl}order-history/[token]`,
   accountOrderHistory: `${accountBaseUrl}order-history`,
-  accountStore: `${accountBaseUrl}store`,
   /*
   Store 
   */
